refactor: Improve naming
